Authentification pour le mode remote + mise en place d'un reverse proxy
