+++ .reverse-proxy/nginx/README.md
@@ -1,0 +1,9 @@
+# Deploiement Docker avec Nginx
+
+> TODO
+
+## Liens utiles
+
+<https://dimuthukasunwp.github.io/Articles/Hosting-multiple-sites-or-applications-using-Docker-and-NGINX-reverse-proxy-with-Letsencrypt-SSL.html>
+
+<https://github.com/SadhvikChirunomula/nginx-docker-multiservice-demo>
