+++ .reverse-proxy/traefik/certs/_wildcard.docker.localhost-key.pem
@@ -1,0 +1,28 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvAIBADANBgkqhkiG9w0BAQEFAASCBKYwggSiAgEAAoIBAQDP2li8N+gAYgg6
+W8ClQmPELJrPsjG2EU/ozt7THlnzV59JzRU+N4iRk1+feq8g/vtdBpoB4bV3GYeN
+KhoUr4SLxLSDzQoE7UPnQe8uARLjxtEbu5rsVhMfDw0NpwVzyYZOF6Yodg1rzdKz
+t1JRsL1c99UFxKaTu8P7R7FfBHHM9hMWzOP3SDMh0keCzO+BO8y8Ev47rWOY3MMz
+mho1Rz6PASwu9/jlyt9sM6leXdH5nU8dWpxhAQu8lZuxcHq6ApSwCvIQegwPfsYQ
+hdlbyu0Dc3JqreKAx3WDhfebmBWKebc45Q2CGY896nz/0EgHrlvnQC/p1oe2NGg6
+7NS8vSfFAgMBAAECggEAbqOWWlIT/6gvIwlWCWHGg1UO7wG2bmjWtWDwwUZFsT9x
+/bHZ3e1nl73vx49hi1nO6VyrO9cuZgBD5EOF8H7hOhjx0bqh26+EVqQFMeLGcFKT
+WoTJWYYT1ZOIajz+rPyXrhfujVkRmdHMl3c8rVk74zHUxaZA2SC4mJg39REDlopr
+h/zVnW2TADNoazJQUX9xr2HguC3a4O/MF3e02ig9OWorT6YVA4t/DIuQds6L8QhF
+1+HHoMa6ij9TKosjckmYkrGXTa3PGVkEA9Apm5lFlToVvf+cSMxQXHctliKKrqwh
+9eSC43ZzA2yTiumwgI6UiRCYRYvlkUdL+vafs9md3QKBgQD6RzpUmMExwMRUdjcO
+6CuyZn0K47vqMgSg3qPDNku0lpSLQfaWx7++ey326JDuq0ou/052F9yMFfvAeJxP
+5U9dm6P4N3PcjDgN+5lVkcqOeXkwiD9saczx70Y9Rc2CdgvJROE1nqnlJ4WjQhzy
+5NlHeCn9/vDoDON9kN5EvYX8wwKBgQDUmtJIQhyYcRI7AUsrLe5u1lVz4Al2iLD7
+yeavRyew+Bz8q/obJoP827jC41ruiWtUEnSfYT8gmqIQob5dcM86ElbF1qrvKyfq
+ypPIfE2pziV5MYHpzlL0hqd+EARSXGQe6i94wO22arrVtWnsbDN3MlppByQnQG3q
+hd/IXBSg1wKBgAyKjS24CdZIZo4k7lVc5POnlxR4nO37W6gEAAjEK7783KytvYGI
+uN7+pSr+NGUBSFJakL/yKicK4vaeFJccUoHKTskdXjzq+dQZi3FyQMgq5hu8eBEi
+rBZjew2O9eFXsPeA8BgTv0ewV036/iBdv1WJe/EyWTgHjU65d6nSJWipAoGAHMZP
+i6QXBrLXCtFkARYYXTv5GNPLtBVQDuQGqtZwks0h7zi/gwk1d/WtDgODfSFwO1K8
+Ig4hYyJKqQGbKD5qbaupvuENBY3Lu0VrtwjL4h2aiMqgkKpyI9JNvQwh6ZvDmUOM
+5UUHZ1cOrnn9Ohhy5sET538l74IS3lj8Lrv3tsUCgYAM/hb33jPAQrk+aIVJFdcg
+XcSGFKwLD/wvg7ujO09QRzlO081Bw/1xtYELp0Crcfjn75XLXQHTkj/qJbzBratr
+Cb22hI2r/SIdoKmMRXZ3r0gLav9YfDxmaia8A74OP3wdx992oRQytr0bUc+DME4U
+hcFul5daeQbZZiNcnSKCkw==
+-----END PRIVATE KEY-----
