+++ .reverse-proxy/traefik/certs/_wildcard.docker.localhost.pem
@@ -1,0 +1,25 @@
+-----BEGIN CERTIFICATE-----
+MIIEJDCCAoygAwIBAgIRAMnjSNxx09hRdO4MD/wNREAwDQYJKoZIhvcNAQELBQAw
+XzEeMBwGA1UEChMVbWtjZXJ0IGRldmVsb3BtZW50IENBMRowGAYDVQQLDBFyb290
+QDdlYjEwOTVkMTM0NjEhMB8GA1UEAwwYbWtjZXJ0IHJvb3RAN2ViMTA5NWQxMzQ2
+MB4XDTI1MDIyNjIyMDMwN1oXDTM1MDIyNjIyMDMwN1owRTEnMCUGA1UEChMebWtj
+ZXJ0IGRldmVsb3BtZW50IGNlcnRpZmljYXRlMRowGAYDVQQLDBFyb290QGFhMTgz
+ODc4MDQyODCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAM/aWLw36ABi
+CDpbwKVCY8Qsms+yMbYRT+jO3tMeWfNXn0nNFT43iJGTX596ryD++10GmgHhtXcZ
+h40qGhSvhIvEtIPNCgTtQ+dB7y4BEuPG0Ru7muxWEx8PDQ2nBXPJhk4Xpih2DWvN
+0rO3UlGwvVz31QXEppO7w/tHsV8Eccz2ExbM4/dIMyHSR4LM74E7zLwS/jutY5jc
+wzOaGjVHPo8BLC73+OXK32wzqV5d0fmdTx1anGEBC7yVm7FweroClLAK8hB6DA9+
+xhCF2VvK7QNzcmqt4oDHdYOF95uYFYp5tzjlDYIZjz3qfP/QSAeuW+dAL+nWh7Y0
+aDrs1Ly9J8UCAwEAAaN1MHMwDgYDVR0PAQH/BAQDAgWgMBMGA1UdJQQMMAoGCCsG
+AQUFBwMBMAwGA1UdEwEB/wQCMAAwHwYDVR0jBBgwFoAUDkB0y2kJgshXmiGdBoCR
+psA7D/swHQYDVR0RBBYwFIISKi5kb2NrZXIubG9jYWxob3N0MA0GCSqGSIb3DQEB
+CwUAA4IBgQCwINJ2C5bpeBiF2i17z4Hj5JHkCfpwKYid0jtvFPxIEAqSCsoBMsAp
+26G77rkWYDbDdO6cE3Hy1UKalbfPW0LTxXj5yp5mek82r8T524DOoC9yxFry6Uzk
+EVoJUnthz9oE5FKTdfU2cg8gi5xdwyhRSVAMnXGqU/2YiGWIO69kInQdJTxZtrs8
+PcfiqgiOcv1Q+KJf0lvmAmC1vElMWeTsUxGpr9OopQ8l4rHEGz7C95AmB9iBce5p
+I+NM/uYwPOmHKJuQtaFdsKWvwr/ar/b/dzLlUj8lCRYZGc3SphHhBX/f8dVEKkyD
+2aCc3YRcO7jupukfCuTKC4eacCdIAKSOFghKMq4aelVok+DwfxaWZF4AIrOVSowm
+R+POQAGyykiUbWSyICM92UiGaX7//xSMeHKk+bftdzyD6xcldFE2dQlT/iugkdsa
+xwW1KWkcSFiSCCCoIZxKp9wkaLW0/gUaH6aqYKT8JBHivymqOACEUjQp9BM/59z9
+sy3evoQsRB4=
+-----END CERTIFICATE-----
