+++ .reverse-proxy/traefik/certs/docker.localhost-key.pem
@@ -1,0 +1,28 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDq0ALY7CgYJBdc
+QpbXevJwHkQOW+3vBbDgIGmNlVisDNvh1w7LAftwZxErVTFrCOmO8r2vOedyNK0o
+0UBZYzcenWcWwfE7/saNnaF6Wd7TKnQucO5HMffbpWOVV6EoZt2zOeol7DNe25tO
+xs36v99lhsnYy2ti4jQAwqr0Dc7kgFF7Gz2oMF1QOl/2rY+xlcZGL4AvutzZ5Kw9
+X5O5Kuu1z9b6cXtK/c5hYkZlEOn5pQ4ngwV//uwtGt6tMe44Isg2oFeC7ttsrVS0
+t2/whaH+zcEtYMGarjj/2QrJQKvqPBv48PDIl5TqI/3JyRR/UtWjvtnpD5IrICRK
+0FDdncnbAgMBAAECggEAJRgBKyV7W/zvl8qJK7cnb2DInYAWQUkj3PQh2JT8MgnB
+/pZLpvRENUiC5NaLfP4mb0eXgYPjNR5vMGCGCtjqjzKMY4L1yeWAlkSNiteBpNgz
+DFcYvTR+lDs3VRNHySse4L5IZ/89SkHk8ffj+DWxY8d8QXw5FaDoWCpoeS9pPJ1f
+PE/Hdqp2tmcu/r5V+IITtE+yIZUF4TVw/kee+OOrfJQ+VQr4IpJnSkQCgrUoD0/g
+lY9ClpGxtL36+JXtBMGJ4N6Eo0LTbNXj14IorxKAdfUHL8Xp7RCGaNWbs1stUI7Q
+9gtLdq8rKiB4rr2DBRyzHvJL3OK2V5bCSB9M2bvHwQKBgQDr3UrRzYm59h2LFdFz
++a/4zsIPm89y+gz2Djh1o/OCuAI1uxiD223kD7/rFvtQoFbE3yHXpHvd2xZw0Cq9
+ZT9ua2RrBgQyvtoZ3O3ctXBlRWkrCOs9P7pfb0qXqPoYbU3xjbd6nNewc/I2uw28
+dOLIhy5q68idZjuRPLNWK1PbYQKBgQD+27sDXve1gUhwvDsRSZBjzEGjFqXrPbLp
+pfOUdrlw2Ta47dCROaTJ1eq9DvCDmiRTrpr8zZVljAKbP5fOBaOc/f00B20zgWhL
+AS48CYZzFnp9vd0bPaPFjTFwD5AsFfBHhrCiQohI67TvW4eL5yJgpglnRR64LiEJ
+FKbIxYLKuwKBgHV1fbYfso/OVgYWtnkYBh9U7+WQBW5Pwjai/FDxdepKMqeJZfr8
+QqyJf1zW1A7h2nQuXo8iVaE1YuaJdE9fZYhaqOZ3V43h54FruXTpUnDMcURCK8wT
+uX+fT7lZKDsT6/X4SkAgbXVtOFkzK/MHDo0obOBITV69M7rvTKwD5b3hAoGBAKnj
+gaO6n7d/ifGUmG9qsXTFSQ/VXh8VTvmPpXLXPcCrG15lMT5NxbrB/auvlMa1bvfV
+kf1pcehyT5574s70VV+mtC4FsVa/Me2lyNM+qkB3wImnkgfBrDVZbJY92aZ5rWWQ
+tJhFb6HkdgWE5zNsjS8OH4CWw+aU4XKzMIbvmsNXAoGATyO9KQwCUq5/BLYOHEFT
+yRBx1842P+84RAvUElHzhIK6qaH0dYXskLMLlf1vMzW3mCte/qtMiFRSyBVMdrUZ
+T7bS7iYm5fw/lEtGZt6R7BrBaRCnrJTJjM5764eewGFQFb+D/GyC30802pu8gxcw
+MEwcrTVLUxqP1MxhseL1mDg=
+-----END PRIVATE KEY-----
