+++ .reverse-proxy/traefik/certs/docker.localhost.pem
@@ -1,0 +1,25 @@
+-----BEGIN CERTIFICATE-----
+MIIEIjCCAoqgAwIBAgIRAOHNgJ7yVcBwXQdtXlokT98wDQYJKoZIhvcNAQELBQAw
+XzEeMBwGA1UEChMVbWtjZXJ0IGRldmVsb3BtZW50IENBMRowGAYDVQQLDBFyb290
+QDdlYjEwOTVkMTM0NjEhMB8GA1UEAwwYbWtjZXJ0IHJvb3RAN2ViMTA5NWQxMzQ2
+MB4XDTI1MDIyNjIyMDMwN1oXDTM1MDIyNjIyMDMwN1owRTEnMCUGA1UEChMebWtj
+ZXJ0IGRldmVsb3BtZW50IGNlcnRpZmljYXRlMRowGAYDVQQLDBFyb290QGFhMTgz
+ODc4MDQyODCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAOrQAtjsKBgk
+F1xCltd68nAeRA5b7e8FsOAgaY2VWKwM2+HXDssB+3BnEStVMWsI6Y7yva8553I0
+rSjRQFljNx6dZxbB8Tv+xo2doXpZ3tMqdC5w7kcx99ulY5VXoShm3bM56iXsM17b
+m07Gzfq/32WGydjLa2LiNADCqvQNzuSAUXsbPagwXVA6X/atj7GVxkYvgC+63Nnk
+rD1fk7kq67XP1vpxe0r9zmFiRmUQ6fmlDieDBX/+7C0a3q0x7jgiyDagV4Lu22yt
+VLS3b/CFof7NwS1gwZquOP/ZCslAq+o8G/jw8MiXlOoj/cnJFH9S1aO+2ekPkisg
+JErQUN2dydsCAwEAAaNzMHEwDgYDVR0PAQH/BAQDAgWgMBMGA1UdJQQMMAoGCCsG
+AQUFBwMBMAwGA1UdEwEB/wQCMAAwHwYDVR0jBBgwFoAUDkB0y2kJgshXmiGdBoCR
+psA7D/swGwYDVR0RBBQwEoIQZG9ja2VyLmxvY2FsaG9zdDANBgkqhkiG9w0BAQsF
+AAOCAYEAQ9yrYUDeSOxRIDl6RexfqZ2r9AdKi2f//FuzQIICIK39TGAmkA4YWmKd
+bRWCwRQ1zxgTO3F4PsrSMW3T2txvtCYdJ/8IXzKSO70FxArBnuuT5E0o/0l3wbOI
+7cwSYxyAaPqw/LRdYFVJV6rATwoSgjXyhCa2AvafDE08W26hGMX9O5pqBjms0WgY
+A5EGPDoiplYGwqgDele4IATanREv+YQFeYUuVJZZ1+l0NIiEXiJF70uHmDc+v9E4
+tID95RuHnqZxOl6Pv2PogAaCRq1aRLQk1puk9EtmMmPpBFKIqAoN4hrE+4BDNCnj
+yXdL2Cbjc3PctGeVMTNAS1ulu88HY7+HUCjs1pRbP4neY18qW4f0FOU/iHXZkT0j
++eXplPsrbmAeXUKZhfu9GiIi72+J1oFadTsqi/ylSqdTiaafiA46cGolW1kIV6VV
+4pWQVd0x5fd3/YlH3iqkFH18ns+bAACbmaqpUAnk5Q9xqyUNOcLoaDKbOcM0XtAo
+DLFdGixO
+-----END CERTIFICATE-----
