+++ .reverse-proxy/traefik/certs/rootCA-key.pem
@@ -1,0 +1,40 @@
+-----BEGIN PRIVATE KEY-----
+MIIG/gIBADANBgkqhkiG9w0BAQEFAASCBugwggbkAgEAAoIBgQDCDBIkhu86yRDl
++YZ9FraYOYX8yj2JxPtwC7Q7kgJ4lz1ZIq9AnZx68fBDT9/MVs9LYv+E0TIHd+DX
+ZjVpZ2uIYHxI1D5UKWuoU83ytFzj5enT1AxpnpvaWHg7pjNjMhR7+8KCP5TnLsir
+BHkke+fNvoxuyuZK7Mk3m083aqDMhwpi8PViGD3JBGTyuO5Ao0qeRJeLNCaf+h+z
+Kx60GAwlU+MoniZM5MDU4fT52G1Bw0eL4WFJPdTnSAzS2N1fX/i+62GMtQ/5M/vy
+MgLJ+0y7yXQ6VFRFlh0U7giD0JcmDo5oZfKx/5lqmWF6THc7216rh8U5XAmHwf49
+5E5MQktiNHG33Wr/5Gzpcgz6QQ3xPaIdsFAKwwuZrxVbfewwjLTTnreD8dF6p8dl
+g5S4CNZUa7MT1DPdDybijqC8EU0ObJoEHfHor2VZyd66LwzWnwnW1lzio5IcFg5v
+tuql1TaI7SLFTQrG6IPXVQNTT/U6AQx+HhT+3wc5gBs9bawvPeMCAwEAAQKCAYBk
+XWvw2p3X0PFdkZ8fCeVtfKsxocUMggbu4A7blNyUVN1VQhxVFnA+Iu77Uc/WV7wm
+KNEyeCUhOb7giivsFaLHyefen3CRhU76pEXsV5+H/H3ho1a8Nl40I/gkGgWQM/Hm
+YsQrLAmvPUJrOggSd8fgYcMu9uEANdKpYToq1orO3rXfkPPKRfGHDC/OXxpT8mGA
+1MdxJ4smyW8l85KQ4bM+h1IByzLSKKsYMLrNk0SHdCKIVg1EnaeF7H09fQ+a6f7O
+2YIi4PDp8cQeheK0QNRo1d4Lx+jGGcmn5R4K58Wrv6BkukelNP4E/PyuNz6l2K/e
+Ib/wX7nIhSt0MN41esWJdcwCYYfMhn9wQQtJsYfWQkwPVaYBDEDGc6uDN1cn0zUq
+SlQ/BQir99f/Ij5LmL0+akzWW9HjiDl654C9+cI/ICtU7WoXj4JwiNeqd6LFlp/w
+q3oJQQZSzMUaKqEQGQY5unoQu/9eDEIPcSAeQNVMHRnwhiyfURbPM8Qw0I+hEJkC
+gcEA2cXtXNxm7Eo4bz1B0XsE2OgTCeEYYNiuGlSutJ926tmHQ25qB3FGqDMjrbSj
+Vz89BdwriDfVVw8r7+1QzHBLYaUJAyvy8Yl1LgWYwI44c5kkWJ5pj/+ifnoEtgHP
+ohAd+/VCppDFjImxT7M1Uy4FnOGVYxqF2nYId1tcaBcs4WWhgSjc0TtSd2MIGkjq
+FhS4Y9mbAGGk1PaKC/e6xAMDCSrgvWBt4tA3BXKomn0K2aDcbPnKJsmwVKvhfc4r
+LaRlAoHBAOQb9+F9oGlXKmyndFJLNqG9YuYTVK4YTAWREQFNgMiLx64JMiLOZxr6
+6xXAr8b6zyHXJXYVxS9PK1IX7ihwwa+9mGPAUq3q+4xDdMGLAbf115eMAsAEEStt
+Hk8zdKkNiljgviZJavXqRx9V6LQQGwAHpASYmUySa3s58iAS7PzB5N/3OiLDH6PO
+YuYFC6PxwBaHb0tDnT1uzUUMfZUHW/swKwCe6BTNbvNwB/f5/KdGDIN7ivlzmT7/
+vxhNzrUApwKBwQDPCyKUuucPl7D38aF17CO7pUkUh3GKAfflCMXg7As/VPkI3xzn
+Xobyu0wchjUVagY2YBDzPUITt1TqkxHSjKDegO2EmnAru6gwzRlhWNXK8bpglpt+
+UGjJHnEEKcyiI/id55wqiFGJQA/X2mXEY2jqsn447qp7/OQntjCwtffxg5J9HcBz
+HJdW8kxt0QzdWW3FYMnBHxR7dLnijvB+Eu779EwfQGK41YS6JTyvRFrgddObuBC9
+t9txZMjJtw3OIxECgcEAtU8RsNZEuVYCL4jOKVcFHnvE7zDwPsyrKEWM6acrtLt9
+xHQkuIusYkKMtqaIULUsDC7Q5NIbACaCPWct8lDlcoVVafwE+53pwSwg0nFltU2f
+xiIQUYUqDIVX8QXqNTDBBTHlkdMLRFmpnFuRcgXXICfyk4CjYLyE19ZQAHRqj+Ch
+xe7g8556OJ4j98aM4QuXbZDKwl0Ezz6Rw89QxjKqRrabh4vG1joPdUlX3B+EVYSV
+D7ET3oCmju+CeZPmt/HDAoHAOq635Su1K64u6XGzAXneGEqvLPTu2/ja2AjjWaTT
+s36anTIT6pHG6spSSpBzXhVUWD4oaodeul+LRpgoHGBfa8PM3Hr17FdhzMrgKbEV
+tV5lcMc5gBRWkiavcyXrPxYvZMw4dIh0UXFz1KHIcb0ESNzUXH9X7wZ9egAszWRQ
+RfjkoA/KNKUal2kS5RupodpdM/tXvDomyZtdmPZ3cYZHtSPmw8ZQBISFIB1mLw8E
+lC4o73QwvlWW2Nwav0g7zwIn
+-----END PRIVATE KEY-----
