+++ .reverse-proxy/traefik/certs/rootCA.pem
@@ -1,0 +1,27 @@
+-----BEGIN CERTIFICATE-----
+MIIEjjCCAvagAwIBAgIRANFwldW0rVdfcYr6n1vCHhEwDQYJKoZIhvcNAQELBQAw
+XzEeMBwGA1UEChMVbWtjZXJ0IGRldmVsb3BtZW50IENBMRowGAYDVQQLDBFyb290
+QDdlYjEwOTVkMTM0NjEhMB8GA1UEAwwYbWtjZXJ0IHJvb3RAN2ViMTA5NWQxMzQ2
+MB4XDTI1MDIyNjE1MjgxNloXDTM1MDIyNjE1MjgxNlowXzEeMBwGA1UEChMVbWtj
+ZXJ0IGRldmVsb3BtZW50IENBMRowGAYDVQQLDBFyb290QDdlYjEwOTVkMTM0NjEh
+MB8GA1UEAwwYbWtjZXJ0IHJvb3RAN2ViMTA5NWQxMzQ2MIIBojANBgkqhkiG9w0B
+AQEFAAOCAY8AMIIBigKCAYEAwgwSJIbvOskQ5fmGfRa2mDmF/Mo9icT7cAu0O5IC
+eJc9WSKvQJ2cevHwQ0/fzFbPS2L/hNEyB3fg12Y1aWdriGB8SNQ+VClrqFPN8rRc
+4+Xp09QMaZ6b2lh4O6YzYzIUe/vCgj+U5y7IqwR5JHvnzb6MbsrmSuzJN5tPN2qg
+zIcKYvD1Yhg9yQRk8rjuQKNKnkSXizQmn/ofsysetBgMJVPjKJ4mTOTA1OH0+dht
+QcNHi+FhST3U50gM0tjdX1/4vuthjLUP+TP78jICyftMu8l0OlRURZYdFO4Ig9CX
+Jg6OaGXysf+Zaplhekx3O9teq4fFOVwJh8H+PeROTEJLYjRxt91q/+Rs6XIM+kEN
+8T2iHbBQCsMLma8VW33sMIy00563g/HReqfHZYOUuAjWVGuzE9Qz3Q8m4o6gvBFN
+DmyaBB3x6K9lWcneui8M1p8J1tZc4qOSHBYOb7bqpdU2iO0ixU0KxuiD11UDU0/1
+OgEMfh4U/t8HOYAbPW2sLz3jAgMBAAGjRTBDMA4GA1UdDwEB/wQEAwICBDASBgNV
+HRMBAf8ECDAGAQH/AgEAMB0GA1UdDgQWBBQOQHTLaQmCyFeaIZ0GgJGmwDsP+zAN
+BgkqhkiG9w0BAQsFAAOCAYEAoYbUWHGT/0aYlj5wnRcaCdWYJN3+ummxjaymxNJM
+fJZeCHqMOIjW8lyHuHitOVffX2zYkKFfwRz+2DJU8O85UpmeysmV+l27ug6/a3UH
+YWDmCLnFjq/OPUeaN16MPk2AMbL/WqANUkHlosPFyDieZEUGgLOod8Ektt8QlMTl
+cnZ864k0CvrhFXZJCledacI/OQ7SOQzdvdMTEjhai/CcWWbqswJLmTcjZkD1fzsi
++Xg586JcGDjqXLeMPgDmZn7mSVTBoB8Ar+UQUZ+T361r0Wy3qcdADLYSZC8K1T+r
+0YzEOHWNw9lgdNCiOi3RtFoULac1o8HWlBiv3kXKHGEdgaS9e+Tj95W6Kw3w4a4M
+Na/8KlXKgNy/vAx9GsvVdmlJMOHpWMBrzhZ4Kq8cj7oT4RZE6QqhBltRxZcLQKyQ
+nejaZlaQvF4SbRr7On4wBnjFp+MwZ/Fp98YhHbqQ40aB2ei2S7ibqUefu0H+rgQi
+uX8gOCMuf5pzn8vV5/VQJNux
+-----END CERTIFICATE-----
