+++ .reverse-proxy/traefik/dynamic.toml
@@ -1,0 +1,11 @@
+[[tls.certificates]]
+  certFile = "/etc/ssl/docker.localhost.pem"
+  keyFile = "/etc/ssl/docker.localhost-key.pem"
+
+[[tls.certificates]]
+  certFile = "/etc/ssl/_wildcard.docker.localhost.pem"
+  keyFile = "/etc/ssl/_wildcard.docker.localhost-key.pem"
+
+[[tls.certificates]]
+  certFile = "/etc/ssl/rootCA.pem"
+  keyFile = "/etc/ssl/rootCA-key.pem"
