+++ src/views/Login.vue
@@ -21,7 +21,7 @@ onMounted(() => {
   var code = urlParams.get('code');
   var session = urlParams.get('session_state');
   var state = urlParams.get('state');
-  var token = urlParams.get('token');
+  var auth = urlParams.get('authentication_failed'); // remote
 
   // Si aucun parametre de session dans l'URL de la route '/login',
   // on redirige vers IAM authentification
@@ -31,13 +31,19 @@ onMounted(() => {
     service.getAccessLogin().then((url) => {
       location.href = url;
     });
+    return;
   }
   // IAM authentification redirige vers la route '/login' aprés validation
-  // Et, elle fournit le 'code' et la 'session' dans l'url
-  // On revient dans l'application !
-  if ((service.mode === "local" && code && session && state) ||
-      (service.mode === "remote" && token)) {
-    router.push({ path : '/' });
+  // Et, elle fournit 
+  // * le 'code' et la 'session' dans l'url pour le mode 'local'
+  // * 'authentication_failed' pour le mode 'remote'
+  // Puis, on revient dans l'application !
+  if (service.mode === "local" && code && session && state) {
+    router.push({ path : '/',  query: { from : 'login', success : 1 } });
+  } else if (service.mode === "remote" && auth !== null) {
+    router.push({ path : '/',  query: { from : 'login', success : auth } });
+  } else {
+    router.push({ path : '/',  query: { from : 'login', success : 0 } });
   }
 });
