+++ src/views/Logout.vue
@@ -18,6 +18,7 @@ onMounted(() => {
   const urlParams = new URLSearchParams(queryString);
   // parametres
   var session = urlParams.get('session_state');
+  var success = urlParams.get('success'); // remote
 
   // Si aucun parametre de session dans l'URL de la route '/logout',
   // on redirige vers IAM de deconnexion
@@ -25,11 +26,16 @@ onMounted(() => {
     service.getAccessLogout().then((url) => {
       location.href = url;
     });
+    return;
   }
   // IAM de deconnexion redirige vers la route '/logout' aprés validation
-  // Et, elle fournit la 'session'
-  if (session) {
-    router.push({ path : '/' });
+  // Et, elle fournit la 'session' en mode 'local' ou 'success' pour le mode 'remote'
+  if (service.mode === "local" && session) {
+    router.push({ path : '/', query : { from: 'lougout', success: 1 } });
+  } else if (service.mode === "remote" && success !== null) {
+    router.push({ path : '/', query : { from: 'lougout', success: parseInt(success, 10) } });
+  } else {
+    router.push({ path : '/', query : { from: 'lougout', success: 0 } });
   }
 });
 </script>
